+++ src/main/java/com/tdep/tadlab/backendservice/repository/JobRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobRepository extends JpaRepository<JobEntity, Long> {
+public interface JobRepository extends JpaRepository<JobEntity, Integer> {
 
 }
